requirements.txt
amqp==5.0.5
appdirs==1.4.4
asgiref==3.3.1
asttokens==2.0.4
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.5
django-celery-beat==2.2.0
django-heroku==0.3.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
executing==0.5.4
filelock==3.0.12
gunicorn==20.0.4
icecream==2.1.0
idna==2.10
kombu==5.0.2
pipenv==2020.11.15
prompt-toolkit==3.0.16
psycopg2==2.8.6
Pygments==2.7.4
PyJWT==2.0.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
vine==5.0.0
virtualenv==20.4.1
virtualenv-clone==0.5.4
wcwidth==0.2.5
whitenoise==5.2.0
